pyproject.toml
[tool.poetry]
name = "plugin-tools"
version = "0.1.0"
description = ""
authors = ["yannemcie <yann@emcie.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mypy = "^1.11.2"
supabase = "^2.9.0"
python-dotenv = "^1.0.1"
parlant = {git = "https://github.com/emcie-co/parlant.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


